fix: Refactoring project
