+++ app/src/main/java/com/example/kookcoching/Adapter/MainFragmentPagerAdapter.kt
@@ -4,17 +4,15 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentPagerAdapter
 import com.example.kookcoching.Fragment.*
-import com.example.kookcoching.Fragment.Information.InformationFragment
 
 class MainFragmentPagerAdapter(fragmentManager: FragmentManager) : FragmentPagerAdapter(fragmentManager, BEHAVIOR_RESUME_ONLY_CURRENT_FRAGMENT) {
 
     override fun getItem(position: Int): Fragment {
         return when(position){
             0 -> HomeFragment()
-            1 -> ShardBoardFragment()
+            1 -> ShareBoardFragment()
             2 -> MajorBoardFragment()
-            3 -> NewsFragment()
-            else -> InformationFragment()
+            else -> ProjectFragment()
         }
     }
     override fun getCount() = 5 // 전체 페이지 수
