--- app/src/main/java/com/example/kookcoching/Fragment/Information/Post.kt
@@ -1,6 +1,0 @@
-package com.example.kookcoching.Fragment.Information
-
-import com.google.firebase.Timestamp
-
-data class Post(var title: String? = null, var content: String? = null, var time: Timestamp){
-}
