+++ app/src/main/java/com/example/kookcoching/Fragment/Share/Post.kt
@@ -1,0 +1,6 @@
+package com.example.kookcoching.Fragment.Share
+
+import com.google.firebase.Timestamp
+
+data class Post(var title: String? = null, var content: String? = null, var time: Timestamp){
+}
