+++ app/src/main/res/font/font.xml
@@ -3,6 +3,11 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
+    <font
+        android:font="@font/ibm"
+        android:fontStyle="normal"
+        android:fontWeight="400"
+        />
 
     <font
         android:font="@font/tmonmonsori"
