--- app/src/main/res/layout/fragment_information.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".Fragment.Information.InformationFragment">
-
-    <androidx.appcompat.widget.Toolbar
-        android:id="@+id/toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="#002060">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="정보 게시판"
-            android:textColor="#FFFFFF"
-            android:textSize="20dp"
-            android:layout_gravity="center"/>
-
-        <Button
-            android:id="@+id/btn_moveToBoard"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="게시글 작성"
-            android:layout_gravity="end"
-            android:textColor="#FFFFFF"/>
-
-    </androidx.appcompat.widget.Toolbar>
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rv_post"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"/>
-
-
-
-</LinearLayout>
