+++ app/src/main/res/layout/home_text_view_scrolling.xml
@@ -11,6 +11,7 @@
         android:textSize="40dp"
         android:textColor="#FFFFFF"
         android:layout_gravity="center"
+        android:fontFamily="@font/tmonmonsori"
         android:text="이거"/>
 
     <TextView
@@ -20,6 +21,7 @@
         android:textSize="40dp"
         android:textColor="#FFFFFF"
         android:layout_gravity="center"
+        android:fontFamily="@font/tmonmonsori"
         android:text="이거"/>
